requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-decouple